bazel: create filegroups for libraries/plugins
Apparently neither cc_library or cc_import targets can be used as input
for pkg_tar rule, so we need to define filegroup targets for them.

As it stands, cc_library targets are used when building the plugin
itself (they should be cc_import rules but there is an issue
for having headers with uncommon extensions ie. .def) and the filegroups
are used for building the VPP based containers.